+++ include/simol/statphys/simulation/Simulation.hpp
@@ -234,7 +234,8 @@ namespace simol
     sampleSystem(dyna, syst);
 
     //---- actual steps -----
-    for (long long int iOfStep  = 0; iOfStep < dyna.nbOfSteps(); ++iOfStep)
+    //2147483547
+    for (long int iOfStep  = 0; iOfStep < dyna.nbOfSteps(); ++iOfStep)
     {
       //--- display progress every time 10% of simulation elapsed ---
       if ((10 * iOfStep) % dyna.nbOfSteps() == 0)
